Refactoring and new default value for ebuku-buku-path
Mostly refactoring of ebuku--search-helper. It is now (hopefully) easier to
grasp. The majority of its functionality has been broken down to these
functions:

- ebuku--bookmarks
- ebuku--insert-bookmark-string

ebuku--bookmarks handles the calling of buku, and retrieves the results. It
returns the result as a list of data. This coulud potentially be used by other
packages, so perhaps the function should be called ebuku-bookmarks (since it
isn't really a private function).

ebuku--insert-bookmark-string handles the formatting and insertion of a bookmark
into a buffer. ebuku--search-helper calls it when inserting all of the results
into the *EBuku* buffer.

ebuku--get-bookmark-count has been removed, since it was no longer used. It
might be the case however that I misunderstand how the original "count" local variable
was used in ebuku--search-helper.